% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-tools.R
\name{mcptool_read}
\alias{mcptool_read}
\title{Read \verb{MCP} tool definition from \code{YAML} file}
\usage{
mcptool_read(path, check = FALSE, resolve = TRUE)
}
\arguments{
\item{path}{Character. Path to the \code{YAML} file containing the
tool definition.}

\item{check}{Logical. Whether to check if the function definition exists;
default is \code{FALSE}}

\item{resolve}{Logical. Whether to try resolving the path, allowing the
package tools to be loaded with \code{package-function_name} syntax; default
if \code{TRUE}; set to \code{FALSE} if \code{path} should be treated
as a file path without resolving.}
}
\value{
List containing the parsed tool definition.
Throws an error if file cannot be read or parsed.
}
\description{
Reads and parses a \verb{MCP} tool definition from a \code{YAML} file.
This is a low-level function used internally by \code{mcptool_load_all}
and \code{mcptool_path}.
}
\examples{

# Read directly from package using \verb{MCP} tool name
mcptool_read("tricobbler-mcp_tool_search_packages")


path <- system.file(
  "mcp", "tools", "tricobbler-mcp_tool_search_packages.yaml",
  package = "tricobbler"
)

tool_def <- mcptool_read(path, resolve = FALSE)

# Check tool name and description
tool_def$name
tool_def$description

}
