% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class002-contract.R
\name{OutputValidator}
\alias{OutputValidator}
\title{OutputValidator}
\usage{
OutputValidator(validate = function() NULL)
}
\arguments{
\item{validate}{Function that validates the stage output; should return TRUE or an error message.}
}
\description{
Encapsulates a validation function for stage output. The function should
accept the stage output (and optionally the input) and return \code{TRUE} on success or a
character error message.
}
